requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23

# Email
Flask-Mail==0.9.1

# HTTP Requests for API
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0

# Security
Flask-Limiter==3.5.0
